ref/split coding on modules
