+++ src/db.ts
@@ -1,0 +1,8 @@
+import { DataBase } from "./types.js";
+
+export const DATA_BASE: DataBase = {
+  usersMap: new Map(),
+  roomsMap: new Map(),
+  socketsMap: new Map(),
+  gamesMap: new Map(),
+};
